Consolidate get_testdata_dir
The only difference between these two functions is the download url,
but the downloaded file is exactly the same (shasum verified).
